🎨️[Meme] : 최근 등록된 짤 검색 기능 부여
- refactor
    - 기존 paging기능을 ID를 통해 검색하도록 적용
- add
    - 일대다 - 다대다 간 Result Aggregation 활용 검색 기능 부여
    - 테스트 진행, 차이점 확인.
        - Assertion을 활용한 테스트 필요

+++ src/main/java/com/memewiki/core/domain/meme/controller/MemeController.java
@@ -29,7 +29,7 @@ public class MemeController {
 
     @GetMapping("/main/{pagingNum}")
     @ApiOperation(value = "최근 등록 밈 출력 API")
-    public ResponseEntity<BaseResponse> findMemesWithPageable(@PathVariable(name = "pagingNum", required = false) Integer pagingNum){
+    public ResponseEntity<BaseResponse> findMemesWithPageable(@PathVariable(name = "pagingNum", required = false) Long pagingNum){
         BaseResponse baseResponse = BaseResponse.of(HttpStatus.OK, "");
         baseResponse.setData(memeService.findMemesWithPageable(pagingNum));
         return ResponseEntity.ok(baseResponse);
@@ -50,5 +50,4 @@ public class MemeController {
         baseResponse.setData(memeFacadeService.findMemesAndHit(memeId));
         return ResponseEntity.ok(baseResponse);
     }
-
 }
